不必要にフック化されていた処理を適切なコンポーネントに転記 #27
